task value has make double

+++ Allgregator.Aux/Converters/ValueMaxToWidthConverter.cs
@@ -7,8 +7,8 @@ namespace Allgregator.Aux.Converters {
     public sealed class ValueMaxToWidthConverter : IMultiValueConverter {
         public object Convert(object[] values, Type targetType, object parameter, CultureInfo culture) {
             if (values.Length > 2) {
-                if (values[0] is int max && max != 0) {
-                    if (values[1] is int value) {
+                if (values[0] is double max && max != 0) {
+                    if (values[1] is double value) {
                         if (values[2] is Grid grid && grid.ColumnDefinitions.Count > 0) {
                             return grid.ActualWidth * value / max;
                         }
